+++ src/pulumi/core-dtap/analytics/databricks/workspaces/engineering.py
@@ -67,6 +67,7 @@ workspace = azure_native.databricks.Workspace(
     ),
     sku=azure_native.databricks.SkuArgs(name="Premium"),
     resource_group_name=resource_groups["infra"].name,
+    tags=platform_config.tags,
 )
 
 outputs["name"] = workspace.name
@@ -179,8 +180,16 @@ datafactory_token = databricks.Token(
 )
 
 # ----------------------------------------------------------------------------------------------------------------------
+# ENGINEERING DATABRICKS WORKSPACE -> CLUSTER TAGS
+# ----------------------------------------------------------------------------------------------------------------------
+
+# https://docs.microsoft.com/en-us/azure/databricks/administration-guide/account-settings/usage-detail-tags-azure#tag-conflict-resolution
+cluster_default_tags = {"x_" + k: v for k, v in platform_config.tags.items()}
+
+# ----------------------------------------------------------------------------------------------------------------------
 # ENGINEERING DATABRICKS WORKSPACE -> CLUSTERS -> SYSTEM CLUSTER
 # ----------------------------------------------------------------------------------------------------------------------
+
 system_cluster_config = workspace_config["clusters"]["system"]
 system_cluster_resource_name = generate_resource_name(
     resource_type="databricks_cluster",
@@ -201,7 +210,7 @@ system_cluster = databricks.Cluster(
         "spark.databricks.delta.preview.enabled": "true",
         **system_cluster_config.get("spark_conf", {}),
     },
-    custom_tags={"ResourceClass": "SingleNode", **platform_config.tags},
+    custom_tags={"ResourceClass": "SingleNode", **cluster_default_tags},
     opts=ResourceOptions(provider=databricks_provider),
 )
 # ----------------------------------------------------------------------------------------------------------------------
@@ -363,7 +372,7 @@ for ref_key, cluster_config in workspace_config.get("clusters", {}).items():
                 "spark.databricks.delta.preview.enabled": "true",
                 **cluster_config.get("spark_conf", {}),
             },
-            custom_tags={"ResourceClass": "SingleNode", **platform_config.tags},
+            custom_tags={"ResourceClass": "SingleNode", **cluster_default_tags},
         )
     # High Concurrency Cluster Type
     elif cluster_config["type"] == "high_concurrency":
@@ -388,7 +397,7 @@ for ref_key, cluster_config in workspace_config.get("clusters", {}).items():
                 "spark.databricks.delta.preview.enabled": "true",
                 **cluster_config.get("spark_conf", {}),
             },
-            custom_tags={"ResourceClass": "Serverless", **platform_config.tags},
+            custom_tags={"ResourceClass": "Serverless", **cluster_default_tags},
         )
 
 # ----------------------------------------------------------------------------------------------------------------------